functions more than 25 lines corrected

+++ src/execution/exec_redir.c
@@ -1,31 +1,45 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   exec_redir.c                                       :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: sreffers <sreffers@student.42madrid.c>     +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2026/01/09 22:49:00 by sreffers          #+#    #+#             */
+/*   Updated: 2026/01/15 22:47:00 by sreffers         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "../../include/minishell.h"
 #include <fcntl.h>
 
-static int	handle_heredoc(char *delimiter)
+static void	setup_tty(int *stdin_backup, int *stdout_backup)
 {
-	int		fd;
-	char	*line;
-	char	*filename = "/tmp/.mini_heredoc_tmp";
-	int		stdin_backup;
-	int		stdout_backup; // <-- NOUVEAU
-	int		fd_tty;
+	int	fd_tty;
 
-	fd = open(filename, O_CREAT | O_WRONLY | O_TRUNC, 0644);
-	if (fd == -1)
-		return (perror("heredoc open"), -1);
-
-	// 1. Sauvegardes
-	stdin_backup = dup(STDIN_FILENO);
-	stdout_backup = dup(STDOUT_FILENO); // <-- NOUVEAU
-
-	// 2. Forcer l'entrée ET la sortie sur le TTY
-	fd_tty = open("/dev/tty", O_RDWR); // <-- O_RDWR (Lecture ET Ecriture)
+	*stdin_backup = dup(STDIN_FILENO);
+	*stdout_backup = dup(STDOUT_FILENO);
+	fd_tty = open("/dev/tty", O_RDWR);
 	if (fd_tty != -1)
 	{
 		dup2(fd_tty, STDIN_FILENO);
-		dup2(fd_tty, STDOUT_FILENO); // <-- Le prompt ira à l'écran !
+		dup2(fd_tty, STDOUT_FILENO);
 		close(fd_tty);
 	}
+}
+
+static void	restore_fds(int stdin_backup, int stdout_backup, int fd)
+{
+	dup2(stdin_backup, STDIN_FILENO);
+	dup2(stdout_backup, STDOUT_FILENO);
+	close(stdin_backup);
+	close(stdout_backup);
+	close(fd);
+}
+
+static int	read_heredoc_lines(int fd, char *delimiter)
+{
+	char	*line;
 
 	while (1)
 	{
@@ -42,17 +56,39 @@ static int	handle_heredoc(char *delimiter)
 		write(fd, "\n", 1);
 		free(line);
 	}
+	return (0);
+}
 
-	// 3. Restaurations
-	dup2(stdin_backup, STDIN_FILENO);
-	dup2(stdout_backup, STDOUT_FILENO); // <-- On remet la redirection fichier
-	close(stdin_backup);
-	close(stdout_backup); // <-- N'oublie pas de close le backup
-	close(fd);
+static int	handle_heredoc(char *delimiter)
+{
+	int		fd;
+	char	*filename;
+	int		stdin_backup;
+	int		stdout_backup;
 
+	filename = "/tmp/.mini_heredoc_tmp";
+	fd = open(filename, O_CREAT | O_WRONLY | O_TRUNC, 0644);
+	if (fd == -1)
+		return (perror("heredoc open"), -1);
+	setup_tty(&stdin_backup, &stdout_backup);
+	read_heredoc_lines(fd, delimiter);
+	restore_fds(stdin_backup, stdout_backup, fd);
 	fd = open(filename, O_RDONLY);
 	unlink(filename);
 	return (fd);
+}
+
+static int	open_redir_fd(t_redirection *redir)
+{
+	if (redir->type == TOKEN_INPUT)
+		return (open(redir->path, O_RDONLY));
+	else if (redir->type == TOKEN_HEREDOC)
+		return (handle_heredoc(redir->path));
+	else if (redir->type == TOKEN_TRUNC)
+		return (open(redir->path, O_WRONLY | O_CREAT | O_TRUNC, 0644));
+	else if (redir->type == TOKEN_APPEND)
+		return (open(redir->path, O_WRONLY | O_CREAT | O_APPEND, 0644));
+	return (-1);
 }
 
 int	handle_redirections(t_list *redirection)
@@ -65,20 +101,9 @@ int	handle_redirections(t_list *redirection)
 	while (tmp)
 	{
 		redir = (t_redirection *)tmp->content;
-		fd = -1;
-		if (redir->type == TOKEN_INPUT)
-			fd = open(redir->path, O_RDONLY);
-		else if (redir->type == TOKEN_HEREDOC)
-			fd = handle_heredoc(redir->path);
-		else if (redir->type == TOKEN_TRUNC)
-			fd = open(redir->path, O_WRONLY | O_CREAT | O_TRUNC, 0644);
-		else if (redir->type == TOKEN_APPEND)
-			fd = open(redir->path, O_WRONLY | O_CREAT | O_APPEND, 0644);
+		fd = open_redir_fd(redir);
 		if (fd == -1)
-		{
-			perror("minishell");
-			return (1);
-		}
+			return (perror("minishell"), 1);
 		if (redir->type == TOKEN_INPUT || redir->type == TOKEN_HEREDOC)
 			dup2(fd, STDIN_FILENO);
 		else
